More restructuring of drawAdventure and some import adjustments.
